Done Room

+++ src/main/java/com/hotel_management/domain/repositories/RoomRepository.java
@@ -3,5 +3,5 @@ package com.hotel_management.domain.repositories;
 import com.hotel_management.domain.entities.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<Room, Long> {
+public interface RoomRepository extends JpaRepository<Room, Integer> {
 }
